requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
